ai_resume_builder_nlp/pyproject.toml
[project]
name = "ai-resume-builder-nlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-adk>=1.18.0",
    "google-generativeai>=0.8.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
